Implementing GcalWeekView to match the mockups.
What to say about this commit.
There's just one FIXME in gcal-week-view.c, about the size/width of the
GtkScrollbar used. That should not be hardcoded.
Also the background color applied to the scrollbar through css-theming might be
better placed inside the code, since the effectis part of the view.
About this view is still missing a way of signaling the current day. We are on
it, yet.