Improved autotools files.
Moved all the style data to a theme directory ala gnome-shell.
Added close-window.svg image.
Added gcal-marshalers.list and rules to Makefile.am to generate the correct
source and header.

Note: The absence of this changes about the marshallers generation was
making imposible to compile the changes about GcalView API. This will complete
that commit.