Added GcalView::create_event_on_current_unit.
this method is to be called whenever someone click on
the "New Event" button of the toolbar. It will instruct the active
view to generate a GcalView::create-event signal with the parametes
of the current unit (a month for GcalYearView, a day for GcalMonthView, etc)
It will allow the view to mark the current unit cell for the event creation.