Reworked drawing of GcalMonthView.
Internals changed. The drawing now requires more calculus, yet, there's a
lot of places where optimizations can be made. Let's wait for performance
issues. I'm still not entirely happy but the work is taking shape.

The drawing code is broken into pieces, so it can be read (not executed
optimally).