date-selector: fix crash on some locales
As stated on the bug below Calendar crashes in some locales because of the
parsing of the string representing the date. Since parsing this string can be
very tricky, we decided to make a simple translatable string with the format it
is needed and th translators can set the proper order according to his locale.

The translatable string goes as "%d/%m/%y" being %d the representation of the
day, %m the representation of the month and, %y, the year. The translators
should place it in the correct order for its locale, for instance en_US, should
place it: "%m/%d/%y"

The validation of the translated string is lacking some strength. It could use
some improving.

Bug: https://bugzilla.redhat.com/show_bug.cgi?id=1226823