month-view: don't show popover when clicking an invalid cell
Because we're using a simple (gint) cast to round y value of
the clicked point, a negative y was rounding to a valid cell
position and showing the popover even when the user clicked
on the blank cells before the actual month view.

Fix that by properly checking if y is negative.

https://bugzilla.gnome.org/show_bug.cgi?id=755910