month-view: reallocate children when date changes
The subscriber does not add components who are already
added to the view - and doesn't need to do it. To overcome
that fact, we have to reallocate the events even when
nothing was added.

Fix that by setting the GcalSubscriberView->children_changed
to TRUE and queueing an allocation right after the date
changes.

https://bugzilla.gnome.org/show_bug.cgi?id=762066