year-view: Change week numbers style
Week numbers in year view was using a small and very light style and
that was making difficult for users to be able to read it.

After a few conversation with gnome-design team, they proposed the best
approach it would be to use the same style as gnome-shell panel does.

The week numbers style was changed to be as close as possible to the
gnome-shell panel.

https://bugzilla.gnome.org/show_bug.cgi?id=759354