window: add can_focus property to GcalMonthView
In the commit, we made the initial value of can_focus for month-view to be set true, which was by default taken to be false.

The problem is that when we open the month-view in gnome calendar, the input focus is not set to the month view which resulted in the failue of navigation keys initially.

To fix this, the default value of the focus is set to true in the window.ui file.

https://bugzilla.gnome.org/show_bug.cgi?id=763160