gcal-window: Dynamic Today's button
The Today button was always sensitive even if the active date was the
current date.

To fix this, I modified the method update_active_date(), using a
conditional to compare the active date and the current date,
if the date is the same then the button is insensitive,
other else the button is sensitive.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=747963