month-view: selection across months
In this commit we change the data type of start_mark_cell and end_mark_cell to GDateTime from gint

Currently, the mark_cell's are gint and while moving across months, they cause a problem as they don't take the new month into consideration but the cell value of the selector.

To fix this, we changed the data type of the mark_cell's and modified all the functions dependant on it.

https://bugzilla.gnome.org/show_bug.cgi?id=763641