year-view: Multiple Columns
Make year view more flexible being able to handle different number of
columns.

View was limited to only show months in a 4x3 grid which was causing
issues when users were trying to use in different screens sizes.

It was added code to implement a dynamic layout, so year view will
adapt the number of months shown per row, depending on the view width.

https://bugzilla.gnome.org/show_bug.cgi?id=753533