month-view: highlight the whole cell of current day
In this commit, we changed the background area of current day.

Presently, the current day is highlighted by the number. This makes difficult for the user to identify it when there are lot of events.

To fix this, we changed the color of background and made background area equal to the cell's area.

https://bugzilla.gnome.org/show_bug.cgi?id=755946