month-view: hide overflow popover on drag and drop
When the user starts a drag operation from the overflow popover,
it should hide itself, otherwise it can block dropping the event
below it. Currently, this is not what happens.

Fix that by making the overflow popover a drag and drop target,
and hide it when it detects drag movement.

https://bugzilla.gnome.org/show_bug.cgi?id=771663