edit-dialog: also set top and bottom margins
It looks like the cursor at the "Notes" section of the
edit dialog has no vertical space to breath, and it's
a little bit annoying as it runs away from the expected
position.

Fix that by adding a small margin at the top and the bottom
of the text view at the "Notes" section.