year-view: Not in sync with month-view
When user changes the date in month-view, there is no sync with
year-view. The other way around works just fine, when user selects a
date in year-view, month-view goes to the same date.

Both views share an "active-date", but year-view doesn't obey that
property in some cases. Month-view also doesn't warn when changing that
property.

This commit fixes both views to keep them in sync, updating and using
"active-date" correctly.

https://bugzilla.gnome.org/show_bug.cgi?id=750478