year-view: Fix showing selected day
In the new layout using GcalDateChooser widget, it was showing all
first day's month selected.

This commit fixes that, implementing a new API in GcalDateChooser widget
to enable/disable showing the selected day.

Year view was also changed to correctly set the date in the
GcalDateChooser widget when it's in the same month of the current date.