edit-dialog: fix date span backward
The problem is that the end date might change to one before the
start date when modifying the start time.

Treat some cases taking into consideration the start date and start
time related to the fact that an event is spanning all day or not.
Separate the handling of the start date and start time.

https://bugzilla.gnome.org/show_bug.cgi?id=772100