date-chooser: Option to don't show other months
Currently GcalDateChooser shows current month's days and the last days
from previous month and first days from the next one with they are on
the same week the current month start or end.

This is fine we are showing only one month, but if you are showing the
whole year, e.g. year-view, that is not so nice.

Therefore, this change GcalDateChooser to have an option to hide other
months' days.