multi-choice: Fix endless timeout
There are two mapped events for each back/forward buttons in this
widget: button-press and button-release.

When you reach the limits like January and December, button-release is
lost, and you don't get it anymore. That caused button_timeout function
to not be disable anymore, so it's kept in loop forever.

This code fix checks in button_timeout function if any of the back/
forward buttons are still mapped. If not, it prevents from calling
button_timout, stopping the loop.

https://bugzilla.gnome.org/show_bug.cgi?id=776459