source-dialog: use case-insensitive sorting in the calendar list
g_strcmp0() is used when comparing the display name of the ESource.
This is not a valid sort, because in this case we will get an order
like "a" < "z" < "B" < "Z", but it should be like "a", "B", "z", "Z".

We will use g_ascii_strcasecmp() to fix it.

https://bugzilla.gnome.org/show_bug.cgi?id=781231