source-dialog: use DBus to spawn Control Center
In a Flatpak'd world, we cannot rely on the fact that
gnome-control-center binary will be installed at the
app's sandbox. But since it's a core app, we can rely
on the fact that it's available outside the sandbox.

To make it work, use DBus rather than raw spawn to
open control center.