event-widget : Add/rework new event sort function
The overflow layout used to display "Other X events" text when
there is no enough space to accomodate more events.

But due to the way the events were sorted, some events are folded to
'Other events' popover even when the cell has enough space to accomodate
the event(s) or when it is not required to be folded.

We rework how the events will be sorted.  First, the events are sorted by their
start dates(starting with the closest).  But if two or more events have same start date
then they are sorted according to their lengths.  And still if they have same start
date and length then they are sorted according to their modified times(the closest first).

https://bugzilla.gnome.org/show_bug.cgi?id=778478