edit-dialog: remove unused spin_button_set_value
The number_of_occurrences was forcefully set to 0,
due to spin_button_set_value being placed incorrectly.
When a recurring event with limit_type as
GCAL_RECURRENCE_COUNT, having some 'x' number of
recurrences was opened in the edit-dialog, it showed
'0' as number_of_occurrences in place of 'x', which
is incorrect.

This patch removes this unnecessary line of code
and leads to correct assignment of number_of_occurrences.

https://bugzilla.gnome.org/show_bug.cgi?id=782755