application: handle command line options right
Commandline arguments are now handled by the primary instance of the
application. And so outputs for --help, and --version are shown there.
Not where the application is invoked.

So hook to handle-local-options so that the arguments are parsed locally,
and sent unparsed ones to primary instance in case it is required.

As a side effect, the arguments passed to --date and --uuid are not updated
in the primary instance on subsequent runs (ie, when the primary instance
is already up and running). So check the commandline arguments as they
arrives rather than saving them to a variable, which won't then be updated.

https://bugzilla.gnome.org/show_bug.cgi?id=773152