project: Fix missing translations
Gettext support in autotools does define ENABLE_NLS macro to 1 in
the autoconf generated configuration file. Due to the migration to
the meson build system, the procedure that defines This macro is
missing.

There is no reason to really disable i18n, so this fix removes the
condition that binds the text domain.

https://bugzilla.gnome.org/show_bug.cgi?id=787815